// QRCode.FEW/ClientApp/src/app/models/custormer_input.ts
export interface custormer_input {
    id: number;
    custormercode: string;
    custormername: string;
    phone: string;
    fax: string;
    taxcode: string;
    email: string;
    vatname: string;
    vataddress: string;
    website: string;
    description: string;
    dtdannguon: string;
    plkhachhang: string;
    pldichvu: string;
    salexyly: string;
    address: string;
    province: string;
    distinct: string;
    ward: string;
    img_url: string;
}